+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Versions follows PEP440 version scheme.
 
 Fixed
 =====
+* admin permissions are required to access Badge API
 * View on site button in the admin panel on the StatusBadge edit page opens correct URL for image
 
 gamma-core 3.2 (2021-10-18)